+++ tests/Converter/StringToArrayTest.php
@@ -18,7 +18,7 @@ class StringToArrayTest extends \PHPUnit_Framework_TestCase
             "toves",
         ];
 
-        $this->assertEquals($expected, $jabberwocky->toArray());
+        $this->assertEquals($expected, $jabberwocky->toHArray(" ")->toArray());
     }
 
     /**
@@ -42,6 +42,8 @@ class StringToArrayTest extends \PHPUnit_Framework_TestCase
         return [
             "Empty String" => [new HString(), null, null, new HArray()],
             "String of integers with null delims" => [new HString("1 2 3 4 5"), null, null, new HArray([1, " ", 2, " " , 3," ", 4, " ", 5])],
+            "String of integers with zero delims" => [new HString("102030405"), 0, null, new HArray(["1", "2", "3", "4", "5"])],
+            "String of integers with string-zero delims" => [new HString("102030405"), "0", null, new HArray(["1", "2", "3", "4", "5"])],
             "String of integers with blank string delims" => [new HString("1 2 3 4 5"), null, null, new HArray([1, " ", 2, " " , 3," ", 4, " ", 5])],
             "String of integers with space delims" => [new HString("1 2 3 4 5"), " ", null, new HArray([1, 2, 3, 4, 5])],
             "String of integers with comma delims" => [new HString("1, 2, 3, 4, 5"), ",", null, new HArray([1, 2, 3, 4, 5])],